requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SocketIO==5.3.6
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
bcrypt>=4.1.2
gunicorn==21.2.0
gevent>=24.10.1
Authlib==1.2.1
requests>=2.31.0
psycopg2-binary>=2.9.9
python-dotenv==1.0.0
